User comment submission and markdown formatting
There is now a front end for submitting comments to articles. An
authentication system (which will probably be reCAPTCHA) has yet to
be implemented though, so right now the site is an easy target for
spam.

Markdown formatting has been added using the hoedown formatter, and
HTML characters are now appropriately escaped when displayed.

A few convenience functions have been added to the cgi module.