Fix comment submission CAPTCHA checking
The reCAPTCHA is no longer invisible, as dealing with several invisible
reCAPTCHAs is asking for trouble. I may go back to the invisible check
someday, but for now I'm sticking to what works.

The deploy script now just sends the source to the deployment target for
compiling on the target machine, as the system openssl dependenct means
the program can no longer be remotely compiled.