Fixed reCAPTCHA by switching to explicit rendering
Google seemingly changed how they handle several captchas on a single
page, so the rendering method was changed to explicit to ensure all
captchas load properly.

This also adds a slight performance boost, because captchas are only
loaded when the reply forms are revealed instead of as soon as possible.