Port to gotham HTTP framework
The website now runs as a HTTP server on its own port to be forwarded
through a reverse proxy instead of as a collection of CGI binaries. This
will massively simplify the process of adding new pages, as well as
making deployment much easier as only one single binary need to be
updated on source code changes.

A consequence of this is that all statically served files have been
moved to the files directory, as they're now being served by
handler/file.rs, which only looks in said directory.

As a temporary fix error handling has been disabled in captcha.rs as
most of it's code has to run in a separate thread to allow for multiple
tokio runtimes.