Cancel button and database streamling
Added a cancel button when creating or editing articles that returns you
to either the home page or the article you were editing.

Also simplified database handlng code by having a module with functions
that handles all database actions. This also increases code reuse. Note
that this remains to be utilized in the auth module.